requirements.txt
fastapi
uvicorn
markitdown
python-multipart
pytesseract
Pillow
